+++ R/print_helpers.R
@@ -4,7 +4,8 @@ header_str <- function(x,
                        sym = "=",
                        min = 2,
                        prefix = character(0),
-                       suffix = character(0)) {
+                       suffix = character(0),
+                       sep_i = " ") {
   full_w <- options("width")$width
   hw1 <- floor(full_w * hw)
   x_len <- nchar(x)
@@ -12,7 +13,7 @@ header_str <- function(x,
   a <- max(min, a)
   out <- paste0(prefix,
                 strrep(sym, a),
-                " ", x, " ",
+                sep_i, x, sep_i,
                 strrep(sym, a),
                 suffix)
   out